Wire --copy-as flag through CLI configuration pipeline
Add copy_as field to ClientConfig and ClientConfigBuilder,
with accessor method and builder method. Wire the flag from
ParsedArgs through ConfigInputs to the client configuration.

The --copy-as=USER[:GROUP] option specifies a user (and optionally
a group) to use for file ownership instead of the user running
rsync. This is useful when running rsync as root but wanting
files owned by a different user.